Move configs to their own file
